Conflicts:
	Assets/Scenes/Dummy_1.unity
	Assets/Scripts/Cinematic.cs
	Library/CurrentLayout.dwlt
	Library/CurrentMaximizeLayout.dwlt
	Library/ScriptAssemblies/Assembly-CSharp.dll
	Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
	Library/assetDatabase3
	Library/expandedItems
	Library/guidmapper
	Library/metadata/00/00000000000000004000000000000000
	Library/metadata/00/0000000000000000b000000000000000
	Library/metadata/8d/8ddfb9394a262ba48b912cf61c0e0c95
	Library/metadata/94/94d5993a3612a024b827a4cb8eb73a13
	Library/metadata/d0/d07a9fada0662db4badef70872ea094e
	Library/metadata/d1/d1effb8c1df12ab4d9754af8ee9fc1d7
	Library/shadercompiler-32bit-1.log
	Library/shadercompiler-64bit-1.log
	YUNPIMIENTO-csharp.sln
	YUNPIMIENTO.sln
	YUNPIMIENTO.userprefs

+++ Assets/Scripts/Game_Master.cs
@@ -118,7 +118,7 @@ public class Game_Master : MonoBehaviour {
 
 		if(slotManager.CheckSlots() == false){
 
-			float result = 0;
+			int result = 0;
 
 			result = slotManager.CheckSequence();
